+++ rqueue-core/src/test/java/com/github/sonus21/rqueue/core/RqueueMessageTest.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2021 Sonu Kumar
+ *  Copyright 2022 Sonu Kumar
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -156,8 +156,8 @@ class RqueueMessageTest extends TestBase {
             .queuedTime(System.nanoTime())
             .period(10000)
             .build();
-    assertTrue(message.isPeriodicTask());
+    assertTrue(message.isPeriodic());
     message.setPeriod(0);
-    assertFalse(message.isPeriodicTask());
+    assertFalse(message.isPeriodic());
   }
 }