+++ rqueue-core/src/test/java/com/github/sonus21/rqueue/listener/JobImplTest.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2021 Sonu Kumar
+ *  Copyright 2022 Sonu Kumar
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -37,6 +37,10 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
+import org.mockito.invocation.InvocationOnMock;
+import org.mockito.junit.jupiter.MockitoSettings;
+import org.mockito.quality.Strictness;
+import org.mockito.stubbing.Answer;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.messaging.converter.MessageConverter;
 import java.time.Duration;
@@ -46,20 +50,26 @@ import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.*;
 
 @CoreUnitTest
+@MockitoSettings(strictness = Strictness.LENIENT)
 class JobImplTest extends TestBase {
 
   private final QueueDetail queueDetail = TestUtils.createQueueDetail("test-queue");
   private final MessageConverter messageConverter = new DefaultRqueueMessageConverter();
-  @Mock RedisConnectionFactory redisConnectionFactory;
-  RqueueMessage rqueueMessage =
+  private final RqueueMessage rqueueMessage =
       RqueueMessageUtils.generateMessage(messageConverter, queueDetail.getName());
   private final MessageMetadata messageMetadata =
       new MessageMetadata(rqueueMessage, MessageStatus.PROCESSING);
-  Object userMessage = "Test Object";
-  @Mock private RqueueMessageMetadataService messageMetadataService;
-  @Mock private RqueueJobDao rqueueJobDao;
-  @Mock private RqueueMessageTemplate rqueueMessageTemplate;
-  @Mock private RqueueLockManager rqueueLockManager;
+  private final Object userMessage = "Test Object";
+  @Mock
+  private RedisConnectionFactory redisConnectionFactory;
+  @Mock
+  private RqueueMessageMetadataService messageMetadataService;
+  @Mock
+  private RqueueJobDao rqueueJobDao;
+  @Mock
+  private RqueueMessageTemplate rqueueMessageTemplate;
+  @Mock
+  private RqueueLockManager rqueueLockManager;
   private RqueueConfig rqueueConfig;
 
   @BeforeEach
@@ -172,6 +182,8 @@ class JobImplTest extends TestBase {
 
   @Test
   void updateMessageStatus() {
+    doReturn(true).when(rqueueLockManager).acquireLock(anyString(), any(), any());
+    doReturn(messageMetadata).when(messageMetadataService).get(messageMetadata.getId());
     JobImpl job = instance();
     job.updateMessageStatus(MessageStatus.PROCESSING);
     assertEquals(MessageStatus.PROCESSING, job.getMessageMetadata().getStatus());
@@ -210,6 +222,8 @@ class JobImplTest extends TestBase {
 
   @Test
   void updateExecutionTime() {
+    doReturn(true).when(rqueueLockManager).acquireLock(anyString(), any(), any());
+    doReturn(messageMetadata).when(messageMetadataService).get(messageMetadata.getId());
     JobImpl job = instance();
     job.execute();
     job.updateExecutionTime(rqueueMessage, MessageStatus.SUCCESSFUL);
@@ -256,6 +270,8 @@ class JobImplTest extends TestBase {
 
   @Test
   void testMessagesAreStoredInMetadataStore() {
+    doReturn(true).when(rqueueLockManager).acquireLock(anyString(), any(), any());
+    doReturn(messageMetadata).when(messageMetadataService).get(messageMetadata.getId());
     JobImpl job = instance();
     job.execute();
     job.checkIn("test..");
@@ -282,5 +298,27 @@ class JobImplTest extends TestBase {
     verify(messageMetadataService, times(1)).delete(messageMetadata.getId());
     FieldUtils.writeField(
         rqueueConfig, "messageDurabilityInTerminalStateInSecond", currentValue, true);
+  }
+
+  @Test
+  void testMessageWasDeletedWhileRunning() throws IllegalAccessException {
+    doReturn(true).when(rqueueLockManager).acquireLock(anyString(), any(), any());
+    MessageMetadata metadata = messageMetadata.toBuilder().deleted(true)
+        .status(MessageStatus.DELETED).build();
+    doReturn(metadata).when(messageMetadataService).get(messageMetadata.getId());
+    JobImpl job = instance();
+    job.execute();
+    job.updateMessageStatus(MessageStatus.FAILED);
+    verify(rqueueJobDao, times(1)).createJob(any(), any());
+    verify(rqueueJobDao, times(2)).save(any(), any());
+    doAnswer(invocation -> {
+      MessageMetadata messageMetadata = invocation.getArgument(0);
+      assertTrue(messageMetadata.isDeleted());
+      assertEquals(MessageStatus.DELETED, messageMetadata.getStatus());
+      return null;
+    }).when(messageMetadataService).save(
+        any(MessageMetadata.class),
+        eq(Duration.ofMinutes(rqueueConfig.getMessageDurabilityInMinute())));
+
   }
 }